data-scraping/requirements.txt
backoff==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
idna==3.7
loguru==0.7.2
python-dateutil==2.9.0.post0
requests==2.31.0
scrapfly-sdk==0.8.16
six==1.16.0
urllib3==2.2.1
